backend/weather_data_analysis/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
uvicorn==0.27.1
requests>=2.31.0
pandas>=2.2.0
numpy==1.26.3
pydantic==2.6.1
pydantic-settings==2.1.0 
aiohttp==3.8.5
aiofiles==2.1.0
celery==5.3.1
httpx==0.25.3
apscheduler==3.10.1
kafka-python>=2.0.0
redis[hiredis]==4.5.0
